requirements.txt
python-nmap==0.7.1
networkx==3.1
matplotlib==3.7.0
